blufor_vests - split configs, add apex chest rigs as backpacks
