+++ addons/blufor_vests/CfgVehicles.hpp
@@ -1,0 +1,5 @@
+class CfgVehicles
+{
+    class Bag_Base;
+    #include "include/tacChestrig/CfgVehicles.hpp"
+};
